fix: correct all methods add: UserController

+++ src/main/java/by/divergenny/traveler/service/impl/UserServiceImpl.java
@@ -1,21 +1,22 @@
 package by.divergenny.traveler.service.impl;
 
-import by.divergenny.traveler.persistence.UserRepository;
 import by.divergenny.traveler.domain.User;
+import by.divergenny.traveler.persistence.UserRepository;
 import by.divergenny.traveler.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
+
     private final UserRepository userRepository;
 
-
     @Override
-    public User findByFirstName(String firstName) {
+    public List<User> findByFirstName(String firstName) {
         return userRepository.findByFirstName(firstName);
     }
 
@@ -31,7 +32,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<User> getAllUsers() {
-        return userRepository.getAllUsers();
+        return userRepository.findAll();
     }
 
     @Override
@@ -45,17 +46,65 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void addUser(User user) {
-        userRepository.addUser(user);
+    public User addUser(User user) {
+        return userRepository.save(user);
     }
 
     @Override
-    public void editUser(String userId, User user) {
-        userRepository.editUser(userId, user);
+    public User editUser(String userId, User userWithNewData) {
+        User userInSystem = userRepository.findById(userId).orElse(null);
+        if (null == userInSystem) {
+            return null;
+        }
+        checkAndSet(userInSystem, userWithNewData);
+        userRepository.save(userInSystem);
+        return userInSystem;
+    }
+
+    public void checkAndSet(User userInSystem, User userWithNewData) {
+        if (null != userWithNewData.getFirstName()
+                && !userWithNewData.getFirstName().isEmpty()
+                && !userInSystem.getFirstName().equals(userWithNewData.getFirstName())
+        ) {
+            userInSystem.setFirstName(userWithNewData.getFirstName());
+        }
+        if (null != userWithNewData.getLastName()
+                && !userWithNewData.getLastName().isEmpty()
+                && !userInSystem.getLastName().equals(userWithNewData.getLastName())
+        ) {
+            userInSystem.setLastName(userWithNewData.getLastName());
+        }
+        if (null != userWithNewData.getDateOfBirth()
+                && !userInSystem.getDateOfBirth().equals(userWithNewData.getDateOfBirth())
+        ) {
+            userInSystem.setDateOfBirth(userWithNewData.getDateOfBirth());
+        }
+        if (null != userWithNewData.getCountry()
+                && !userWithNewData.getCountry().isEmpty()
+                && !userInSystem.getCountry().equals(userWithNewData.getCountry())
+        ) {
+            userInSystem.setCountry(userWithNewData.getCountry());
+        }
+        if (null != userWithNewData.getCity()
+                && !userWithNewData.getCity().isEmpty()
+                && !userInSystem.getCity().equals(userWithNewData.getCity())
+        ) {
+            userInSystem.setCity(userWithNewData.getCity());
+        }
+        if (null != userWithNewData.getAbout()
+                && !userInSystem.getAbout().equals(userWithNewData.getAbout())
+        ) {
+            userInSystem.setAbout(userWithNewData.getAbout());
+        }
     }
 
     @Override
-    public void delete(String userId) {
+    public User deleteById(String userId) {
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isEmpty()) {
+            return null;
+        }
         userRepository.deleteUserById(userId);
+        return user.get();
     }
 }
